Fix undefined share functions
Added definitions for the `shareOnTwitter` and `shareOnLinkedIn` functions in `BlogPostDetail.tsx` to resolve runtime and TypeScript errors related to undefined references.
[skip gpt_engineer]
